Changed type of version from int to short in these classes.
git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@789 0b15c3b6-4604-4ca4-ad59-2e1af984e93c

+++ src/edu/umass/cs/gns/nsdesign/Config.java
@@ -16,7 +16,7 @@ import java.util.HashMap;
 public class Config {
 
   /** First version number for a GUID. */
-  public static final int FIRST_VERSION = 1;
+  public static final short FIRST_VERSION = 1;
 
   public static final String NO_COORDINATOR_STATE_MARKER = "NoCoordinatorState";
